mcu_interrupt: optimize pending interrupts
- Extract interrupt level and vector translation to a separate function
  to avoid control flow difficult for the optimizer. In particular,
  removing `continue` and using positive logic provided an ~8% speedup
  on gcc and ~5% on clang.
- Avoid checking all possible pending exceptions and interrupts.
  Instead, use a bitset that allows iterating over the currently pending
  interrupts only. In practice this makes the pending loops run fewer
  than 5 iterations instead of 16 and 20 respectively in the worst
  cases.

With these changes combined, render times are faster across the board
(Release, LTO enabled):

gcc:   1.16x faster
clang: 1.10x faster
msvc:  1.12x faster

+++ src/backend/mcu_interrupt.h
@@ -70,5 +70,5 @@ enum MCU_Exception_Source : int8_t {
 
 void MCU_Interrupt_SetRequest(mcu_t& mcu, MCU_Interrupt_Source interrupt, bool value);
 void MCU_Interrupt_Exception(mcu_t& mcu, MCU_Exception_Source exception);
-void MCU_Interrupt_TRAPA(mcu_t& mcu, uint32_t vector);
+void MCU_Interrupt_TRAPA(mcu_t& mcu, uint8_t vector);
 void MCU_Interrupt_Handle(mcu_t& mcu);
